ТЗ-12. Переход на БД. Коммит для перехода на другой HDD...

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -32,7 +32,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}/genres")
-    public List<Genre> getGenresOfFilm(@PathVariable Integer id) {
+    public List<Genre> getGenresOfFilm(@PathVariable Long id) {
         log.debug("Запрошены жанры фильма с ID = {}", id);
         return service.getGenresOfFilm(id);
     }
