+++ src/main/java/ru/yandex/practicum/filmorate/repository/GenreRepository.java
@@ -12,7 +12,7 @@ public interface GenreRepository {
 
     Optional<Genre> get(Integer id);
 
-    List<Integer> getGenresIdsOfFilm(Integer id);
+    List<Integer> getGenresIdsOfFilm(Long id);
 
-    List<Genre> getGenresOfFilm(Integer filmId);
+    List<Genre> getGenresOfFilm(Long filmId);
 }
